Move missed endpoints
